More categorization of wasm_backend test failures
