+++ tests/core/test_sizeof.cpp
@@ -24,5 +24,6 @@ int main(int argc, const char *argv[]) {
 
   printf("*%d,%d,%d,%d,%d,%d*\n", as[0].x, as[0].y, as[1].x, as[1].y, as[2].x,
          as[2].y);
+
   return 0;
 }
