+++ tests/core/test_sizeof.out
@@ -1,1 +1,3 @@
-*2,2,5,8,8***8,8,5,8,8***7,2,6,990,7,2*
+*2,2,5,8,8*
+*8,8,5,8,8*
+*7,2,6,990,7,2*
